[feature] simple pre-flight auth checks (#366)
[feature] simple pre-flight auth checksBefore running any terraform commands, run a quick authentication check.

This won't catch every problem (like authorization failures), but in the cases where a profile is not configured at all, this should give a better error message than terraform does.

Thanks to @kislyuk  for pushing for a better experience here.

Note that his assumes the aws cli is installed. Do we think that is a fair assumption?

### Test Plan
* make test
* running a prerelease on shared-infra

### References
(Optional) Additional links to provide more context.